requirements.txt
# Potato Disease Classification - Complete Requirements
# Install all dependencies with: pip install -r requirements.txt

# Core Machine Learning
tensorflow>=2.10,<3
numpy>=1.21.0
pillow>=9.0.0

# Web Framework
fastapi>=0.114.0
uvicorn[standard]>=0.18.0

# Frontend
streamlit>=1.35.0
requests>=2.28.0

# Development Tools
jupyter>=1.0.0
notebook>=6.4.0

# Optional: For better performance

# Data Processing
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Image Processing
opencv-python>=4.5.0
scikit-image>=0.19.0

# Model Evaluation
scikit-learn>=1.0.0

# API Testing
httpx>=0.23.0
pytest>=7.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=8.0.0
